Move cluster operator controller to a separate package
