+++ pkg/controllers/common_consts.go
@@ -2,9 +2,6 @@ package controllers
 
 const (
 	DefaultManagedNamespace = "openshift-cluster-api"
-
-	infrastructureResourceName = "cluster"
-	clusterOperatorName        = "cluster-api"
-	operatorVersionKey         = "operator"
-	externalFeatureGateName    = "cluster"
+	OperatorVersionKey      = "operator"
+	ClusterOperatorName     = "cluster-api"
 )
