--- pkg/controllers/watch_predicates.go
@@ -1,80 +1,0 @@
-package controllers
-
-import (
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-
-	configv1 "github.com/openshift/api/config/v1"
-)
-
-func toClusterOperator(client.Object) []reconcile.Request {
-	return []reconcile.Request{{
-		NamespacedName: client.ObjectKey{Name: clusterOperatorName},
-	}}
-}
-
-func toUserDataSecret(client.Object) []reconcile.Request {
-	return []reconcile.Request{{
-		NamespacedName: client.ObjectKey{Name: managedUserDataSecretName, Namespace: SecretSourceNamespace},
-	}}
-}
-
-func clusterOperatorPredicates() predicate.Funcs {
-	isClusterOperatorCluster := func(obj runtime.Object) bool {
-		co, ok := obj.(*configv1.ClusterOperator)
-		return ok && co.GetName() == clusterOperatorName
-	}
-
-	return predicate.Funcs{
-		CreateFunc:  func(e event.CreateEvent) bool { return isClusterOperatorCluster(e.Object) },
-		UpdateFunc:  func(e event.UpdateEvent) bool { return isClusterOperatorCluster(e.ObjectNew) },
-		GenericFunc: func(e event.GenericEvent) bool { return isClusterOperatorCluster(e.Object) },
-		DeleteFunc:  func(e event.DeleteEvent) bool { return isClusterOperatorCluster(e.Object) },
-	}
-}
-
-func infrastructurePredicates() predicate.Funcs {
-	isInfrastructureCluster := func(obj runtime.Object) bool {
-		infra, ok := obj.(*configv1.Infrastructure)
-		return ok && infra.GetName() == infrastructureResourceName
-	}
-
-	return predicate.Funcs{
-		CreateFunc:  func(e event.CreateEvent) bool { return isInfrastructureCluster(e.Object) },
-		UpdateFunc:  func(e event.UpdateEvent) bool { return isInfrastructureCluster(e.ObjectNew) },
-		GenericFunc: func(e event.GenericEvent) bool { return isInfrastructureCluster(e.Object) },
-		DeleteFunc:  func(e event.DeleteEvent) bool { return isInfrastructureCluster(e.Object) },
-	}
-}
-
-func featureGatePredicates() predicate.Funcs {
-	isFeatureGateCluster := func(obj runtime.Object) bool {
-		featureGate, ok := obj.(*configv1.FeatureGate)
-		return ok && featureGate.GetName() == externalFeatureGateName
-	}
-
-	return predicate.Funcs{
-		CreateFunc:  func(e event.CreateEvent) bool { return isFeatureGateCluster(e.Object) },
-		UpdateFunc:  func(e event.UpdateEvent) bool { return isFeatureGateCluster(e.ObjectNew) },
-		GenericFunc: func(e event.GenericEvent) bool { return isFeatureGateCluster(e.Object) },
-		DeleteFunc:  func(e event.DeleteEvent) bool { return isFeatureGateCluster(e.Object) },
-	}
-}
-
-func userDataSecretPredicate(targetNamespace string) predicate.Funcs {
-	isOwnedUserDataSecret := func(obj runtime.Object) bool {
-		secret, ok := obj.(*corev1.Secret)
-		return ok && secret.GetNamespace() == targetNamespace && secret.GetName() == managedUserDataSecretName
-	}
-
-	return predicate.Funcs{
-		CreateFunc:  func(e event.CreateEvent) bool { return isOwnedUserDataSecret(e.Object) },
-		UpdateFunc:  func(e event.UpdateEvent) bool { return isOwnedUserDataSecret(e.ObjectNew) },
-		DeleteFunc:  func(e event.DeleteEvent) bool { return isOwnedUserDataSecret(e.Object) },
-		GenericFunc: func(e event.GenericEvent) bool { return isOwnedUserDataSecret(e.Object) },
-	}
-}
